update review changes

+++ src/Http/Controllers/Ledgers/LedgerController.php
@@ -6,11 +6,14 @@ use Kanexy\Cms\Controllers\Controller;
 use Kanexy\Cms\Setting\Models\Setting;
 use Kanexy\LedgerFoundation\Model\Ledger;
 use Kanexy\LedgerFoundation\Http\Requests\StoreLedgerRequest;
+use Kanexy\LedgerFoundation\Policies\LedgerPolicy;
 
 class LedgerController extends Controller
 {
     public function index()
     {
+        $this->authorize(LedgerPolicy::VIEW, Ledger::class);
+
         $ledgers = Ledger::paginate();
 
         return view("ledger-foundation::ledger.index", compact('ledgers'));
@@ -18,9 +21,12 @@ class LedgerController extends Controller
 
     public function create()
     {
+        $this->authorize(LedgerPolicy::CREATE, Ledger::class);
+
         $asset_types = Setting::getValue('asset_types',[]);
         $asset_classes = Setting::getValue('asset_classes',[]);
         $commodity_types = Setting::getValue('commodity_types',[]);
+
         return view("ledger-foundation::ledger.create", compact('asset_types','asset_classes','commodity_types'));
     }
 
@@ -39,10 +45,13 @@ class LedgerController extends Controller
 
     public function edit($id)
     {
+        $this->authorize(LedgerPolicy::EDIT, Ledger::class);
+
         $ledger = Ledger::findOrFail($id);
         $asset_types = Setting::getValue('asset_types',[]);
         $asset_classes = Setting::getValue('asset_classes',[]);
         $commodity_types = Setting::getValue('commodity_types',[]);
+
         return view("ledger-foundation::ledger.edit", compact('ledger','asset_types','asset_classes','commodity_types'));
     }
 
@@ -65,6 +74,8 @@ class LedgerController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize(LedgerPolicy::DELETE, Ledger::class);
+
         $ledger = Ledger::findOrFail($id);
         $ledger->delete();
